Fixed DELETE endpoint for behaviors when no behaviors are present on the server.

+++ src/main/java/it/uniroma1/dis/jacov2/server/BehaviorsResource.java
@@ -112,9 +112,13 @@ public class BehaviorsResource {
 	public Response clearBehaviors(@PathParam("client_id") String clientId) {
 		File behaviorsRootPath = new File(rootPath + File.separator + clientId + File.separator + "Behaviors");
 		File[] behaviorsDirectories = behaviorsRootPath.listFiles();
+		if (behaviorsDirectories == null)
+			return Response.ok().build();
 		
 		for (File behaviorDirectory : behaviorsDirectories) {
 			File[] behaviorFiles = behaviorDirectory.listFiles();
+			if (behaviorFiles == null)
+				continue;
 			
 			for (File behaviorFile : behaviorFiles) {
 				behaviorFile.delete();
